Scope OpenCV DLL install and share video comparison
