+++ tests/python_tests/samples/test_denoising_process.py
@@ -2,38 +2,11 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import sys
-from pathlib import Path
 
-import cv2
-import numpy as np
 import pytest
 
 from conftest import SAMPLES_PY_DIR, SAMPLES_CPP_DIR
-from test_utils import run_sample
-
-# Must match the constants hardcoded in the denoising_process samples.
-NUM_INFERENCE_STEPS = 20
-IMAGE_HEIGHT = 512
-IMAGE_WIDTH = 512
-CHANNELS = 3
-EXPECTED_VIDEO_SHAPE = (NUM_INFERENCE_STEPS, IMAGE_HEIGHT, IMAGE_WIDTH, CHANNELS)
-
-
-def _read_video_frames(path: Path) -> np.ndarray:
-    cap = cv2.VideoCapture(str(path))
-    try:
-        assert cap.isOpened(), f"OpenCV failed to open {path}"
-        frames = []
-        while True:
-            ok, frame = cap.read()
-            if not ok:
-                break
-            frames.append(frame)
-    finally:
-        cap.release()
-
-    assert frames, f"Could not read any frames from {path}"
-    return np.stack(frames)
+from test_utils import run_sample, compare_videos
 
 
 class TestDenoisingProcess:
@@ -66,18 +39,9 @@ class TestDenoisingProcess:
         assert py_video.exists(), f"Python sample did not produce {py_video}"
         assert cpp_video.exists(), f"C++ sample did not produce {cpp_video}"
 
-        # Both samples write MJPG AVIs via cv::CAP_OPENCV_MJPEG, so the OpenCV-built-in
-        # encoder produces reproducible frames independent of any system FFmpeg/GStreamer.
-        py_frames = _read_video_frames(py_video)
-        cpp_frames = _read_video_frames(cpp_video)
-        assert py_frames.shape == EXPECTED_VIDEO_SHAPE, (
-            f"Python AVI shape {py_frames.shape}, expected {EXPECTED_VIDEO_SHAPE}"
-        )
-        assert cpp_frames.shape == EXPECTED_VIDEO_SHAPE, (
-            f"C++ AVI shape {cpp_frames.shape}, expected {EXPECTED_VIDEO_SHAPE}"
-        )
-        assert py_frames.dtype == np.uint8
-        assert cpp_frames.dtype == np.uint8
-        assert np.array_equal(py_frames, cpp_frames), (
+        # Both samples write MJPG AVIs via cv::CAP_OPENCV_MJPEG, so the OpenCV
+        # built-in encoder produces reproducible frames independent of any
+        # system FFmpeg/GStreamer backend.
+        assert compare_videos(py_video, cpp_video), (
             "Decoded denoising-process video frames produced by the Python and C++ samples are not exactly equal"
         )
